chore: view 클래스들에 대한 view 패키지 생성 및 메소드명 변경
