+++ src/main/java/racingcarwinner/view/OutputView.java
@@ -1,4 +1,6 @@
-package racingcarwinner;
+package racingcarwinner.view;
+
+import racingcarwinner.Car;
 
 import java.util.List;
 
@@ -11,7 +13,7 @@ public class OutputView {
         System.out.println();
     }
 
-    public void getResult(String winner) {
+    public void result(String winner) {
         System.out.println(winner + "가 최종 우승했습니다.");
     }
 }
